+++ packages/node-core/src/index.ts
@@ -16,3 +16,4 @@ export * from './utils';
 export * from './indexer';
 export * from './subcommands';
 export * from './yargs';
+export * from './admin';